feat(v8): add command to sync V8 deps
Add `git node v8 deps` to (re)sync V8 dependencies from V8's `DEPS`
file.
