+++ lib/update-v8/index.js
@@ -5,6 +5,7 @@ import updateVersionNumbers from './updateVersionNumbers.js';
 import commitUpdate from './commitUpdate.js';
 import majorUpdate from './majorUpdate.js';
 import minorUpdate from './minorUpdate.js';
+import updateDeps from './deps.js';
 import updateV8Clone from './updateV8Clone.js';
 
 export function major(options) {
@@ -29,6 +30,14 @@ export async function backport(options) {
   options.gpgSign = options.gpgSign ? ['-S'] : [];
   const tasks = new Listr(
     [updateV8Clone(), doBackport(options)],
+    getOptions(options)
+  );
+  return tasks.run(options);
+};
+
+export async function deps(options) {
+  const tasks = new Listr(
+    [updateDeps()],
     getOptions(options)
   );
   return tasks.run(options);
