+++ lib/update-v8/majorUpdate.js
@@ -1,17 +1,12 @@
 import path from 'node:path';
-import { promises as fs } from 'node:fs';
 
 import { getCurrentV8Version } from './common.js';
+import updateV8Deps from './deps.js';
 import {
-  getNodeV8Version,
-  filterForVersion,
-  addToGitignore,
-  replaceGitignore,
   removeDirectory,
   isVersionString
 } from './util.js';
 import applyNodeChanges from './applyNodeChanges.js';
-import { chromiumGit, v8Deps } from './constants.js';
 import { forceRunAsync } from '../run.js';
 
 export default function majorUpdate() {
@@ -105,66 +100,4 @@ function addDepsV8() {
       spawnArgs: { cwd: ctx.nodeDir, stdio: 'ignore' }
     })
   };
-}
-
-function updateV8Deps() {
-  return {
-    title: 'Update V8 DEPS',
-    task: async(ctx) => {
-      const newV8Version = await getNodeV8Version(ctx.nodeDir);
-      const repoPrefix = newV8Version.majorMinor >= 86 ? '' : 'v8/';
-      const deps = filterForVersion(v8Deps.map((v8Dep) => ({
-        ...v8Dep,
-        repo: `${repoPrefix}${v8Dep.repo}`,
-        path: v8Dep.repo
-      })), newV8Version);
-      if (deps.length === 0) return;
-      for (const dep of deps) {
-        if (dep.gitignore) {
-          if (typeof dep.gitignore === 'string') {
-            await addToGitignore(ctx.nodeDir, dep.gitignore);
-          } else {
-            await replaceGitignore(ctx.nodeDir, dep.gitignore);
-          }
-        }
-        const [repo, commit] = await readDeps(ctx.nodeDir, dep.repo);
-        const thePath = path.join(ctx.nodeDir, 'deps/v8', dep.path);
-        await fetchFromGit(thePath, repo, commit);
-      }
-    }
-  };
-}
-
-async function readDeps(nodeDir, depName) {
-  const depsStr = await fs.readFile(path.join(nodeDir, 'deps/v8/DEPS'), 'utf8');
-  const start = depsStr.indexOf('deps = {');
-  const end = depsStr.indexOf('\n}', start) + 2;
-  const depsDeclaration = depsStr.substring(start, end).replace(/^ *#.*/gm, '');
-  const Var = () => chromiumGit; // eslint-disable-line no-unused-vars
-  let deps;
-  eval(depsDeclaration); // eslint-disable-line no-eval
-  const dep = deps[depName];
-  if (!dep) {
-    throw new Error(`V8 dep "${depName}" not found in DEPS file`);
-  }
-  if (typeof dep === 'object') {
-    return dep.url.split('@');
-  }
-  return dep.split('@');
-}
-
-async function fetchFromGit(cwd, repo, commit) {
-  await fs.mkdir(cwd, { recursive: true });
-  await exec('init');
-  await exec('remote', 'add', 'origin', repo);
-  await exec('fetch', 'origin', commit);
-  await exec('reset', '--hard', 'FETCH_HEAD');
-  await removeDirectory(path.join(cwd, '.git'));
-
-  function exec(...options) {
-    return forceRunAsync('git', options, {
-      ignoreFailure: false,
-      spawnArgs: { cwd, stdio: 'ignore' }
-    });
-  }
 }
