+++ lib/update-v8/util.js
@@ -37,13 +37,18 @@ export function filterForVersion(list, version) {
 
 export async function addToGitignore(nodeDir, value) {
   const gitignorePath = path.join(nodeDir, 'deps/v8/.gitignore');
-  await fs.appendFile(gitignorePath, `${value}\n`);
+  const gitignore = await fs.readFile(gitignorePath, 'utf8');
+  if (!gitignore.includes(value)) {
+    await fs.appendFile(gitignorePath, `${value}\n`);
+  }
 }
 
 export async function replaceGitignore(nodeDir, options) {
   const gitignorePath = path.join(nodeDir, 'deps/v8/.gitignore');
   let gitignore = await fs.readFile(gitignorePath, 'utf8');
-  gitignore = gitignore.replace(options.match, options.replace);
+  if (!gitignore.includes(options.replace)) {
+    gitignore = gitignore.replace(options.match, options.replace);
+  }
   await fs.writeFile(gitignorePath, gitignore);
 }
